Add dynamic partner title to product selection screen.
