+++ src/components/Stratosphere/RedirectBanner.tsx
@@ -3,39 +3,22 @@ import { Alert, AlertActionCloseButton, Text, TextContent } from '@patternfly/re
 import { useSelector } from 'react-redux';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { ReduxState } from '../../redux/store';
-
-// TODO: Figure out what param chrome should expect
-export const AWS_BANNER_NAME = 'from-aws';
-export const AZURE_BANNER_NAME = 'from-azure';
-export const GCP_BANNER_NAME = 'from-gcp';
-
-const partnerMapper: { [partner: string]: string } = {
-  [AWS_BANNER_NAME]: 'AWS',
-  [AZURE_BANNER_NAME]: 'Microsoft Azure',
-  [GCP_BANNER_NAME]: 'Google Cloud',
-};
-const possibleParams = [AWS_BANNER_NAME, AZURE_BANNER_NAME, GCP_BANNER_NAME];
-
-const hasPartner = (params: URLSearchParams) => possibleParams.find((param) => params.has(param));
+import useMarketplacePartner from '../../hooks/useMarketplacePartner';
 
 const RedirectBanner = () => {
-  const { pathname, search, hash, state } = useLocation();
+  const { pathname, hash, state } = useLocation();
+  const { partnerId, partner, removePartnerParam } = useMarketplacePartner();
   const navigate = useNavigate();
-  const params = new URLSearchParams(search);
-  const isVisible = hasPartner(params);
-  const partner = isVisible ? partnerMapper[isVisible] : null;
   const product = useSelector<ReduxState, string | undefined>((state) => state.chrome.activeProduct);
 
   const handleClose = () => {
     // remove only the flag search param
-    params.delete(AWS_BANNER_NAME);
-    params.delete(AZURE_BANNER_NAME);
-    params.delete(GCP_BANNER_NAME);
+    const clearedParams = removePartnerParam();
     // only change the search params
     navigate(
       {
         pathname,
-        search: params.toString(),
+        search: clearedParams.toString(),
         hash,
       },
       {
@@ -45,7 +28,7 @@ const RedirectBanner = () => {
     );
   };
   // show the banner only if correct search param exists
-  return isVisible ? (
+  return partnerId ? (
     <Alert
       actionClose={<AlertActionCloseButton data-testid="stratosphere-banner-close" onClose={handleClose} />}
       isInline
